Complete first chapter solutions (/attempts)
These exercises were completed some time ago, but I had always meant to
do the whole thing and put it in git for the purposes of learning to put
things in git.

Presently iffy on file naming. A lack of description smells like a Bad
Thing for sure, but I would be hard pressed to come up with good names
for each one, hence the readme and comments.